+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `markX()`
   - `markY()`
   - `markPad()`
+- Glide FluentUrlBuilder instances can now be cloned using the `FluentUrlBuilder::clone()`
 
 ### Fixed
 - Setting an encoding option without an explicit quality value clears the quality parameter